refactor: Restruct files & directory for `SearchForm`
